requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.2.0
dj-database-url==0.5.0
Django==4.2
django-filter==23.3
django-js-asset==1.2.2
django-paypal==2.0
gunicorn==21.2.0
idna==3.4
packaging==23.1
phonenumbers==8.12.29
psycopg2==2.9.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.31.0
six==1.15.0
sqlparse==0.4.2
tzdata==2023.3
urllib3==1.25.10
vboxapi==1.0
whitenoise==5.3.0
